requirements.txt
groq>=0.4.0
pyaudio>=0.2.13
speechrecognition>=3.10.0
pyttsx3>=2.90
python-dotenv>=1.0.0
